Add files via upload
This file contains the word by word English translation. Mostly follows Saheeh Int'l. It has notes column that mentions changes and future changes. Note that the words are different. Also fixed transliteration of Names of Prophets and Messengers to IJMES. Note Ibrāhīm is spelled Ibrāhim in Sura 2 as this is a feature of the Quran Printing Complex text, which is used throughout Egypt and lands east. Note that is not in the other readings or the ḥafṣ muṣḥafs that are produced in Turkey and Subcontinent. Other countries that follow use ḥafṣ may or may not have this depending on how they sourced the text.