Add files via upload
Rationalize the text provided by Quran Printing Complex in Medina (QPC) with the one from Tanzil
The total word counts are as follows:
QPC    77,430
Tanzil, 77,800

The reason for the difference that some words that Tanzil text considers separate 

For example: 
Yā + word in QPC are one word and in Tanzil 2 words see 2:21 word 1 and 2 يا أيها "O you people" = ياأيها

one case 3 words in Tanzil and one word in QPC, see 20:94 word 2,3 and 4 يا ابن أم becomes ياابنأم  

So the Tanzil text and where the words are combined in QPC it was combined in the Tanzil text. 
included here is:
Tanzil text with Harkat 
Tanzil without harakat
English Translation, 
Letter shape transliteration (Archigraphemic) following the system devised by Thomas Milo.